Code against the interface / move builder to interface.
