+++ library/src/main/java/com/novoda/storagepathfinder/AndroidDeviceStorageInspector.java
@@ -10,23 +10,17 @@ import java.util.List;
 import java.util.Set;
 import java.util.TreeSet;
 
-/**
- * The inspector that can give all storage roots. both Primary and Secondary. (Internal External and External External)
- * Lots of sanitisation has to go on in these for null checks / empty lists etc.
- * <p>
- * Not it uses a set of secondary storage inspectors to find it in anyway it can.
- */
-public class AndroidDeviceStorageInspector implements DeviceStorageInspector {
+class AndroidDeviceStorageInspector implements DeviceStorageInspector {
 
     private final PrimaryDeviceStorageInspector primaryStorageInspector;
     private final List<SecondaryDeviceStorageInspector> secondaryStorageInspectors = new ArrayList<>();
     private final FileSystem fileSystem;
 
-    public AndroidDeviceStorageInspector(
-        Context context,
-        FileSystem fileSystem,
-        DeviceFeatures deviceFeatures,
-        AndroidSystem androidSystem
+    AndroidDeviceStorageInspector(
+            Context context,
+            FileSystem fileSystem,
+            DeviceFeatures deviceFeatures,
+            AndroidSystem androidSystem
     ) {
         this.fileSystem = fileSystem;
         ExternalStorageDirectories externalStorageDirectories = fileSystem.getCommonDirectories();
@@ -101,62 +95,5 @@ public class AndroidDeviceStorageInspector implements DeviceStorageInspector {
             }
             return lhs.hashCode() - rhs.hashCode();
         };
-    }
-
-    public static Builder builder(Context context) {
-        return Builder.newInstance(context);
-    }
-
-    public static final class Builder {
-
-        private final Context context;
-        private FileSystem fileSystem;
-        private DeviceFeatures deviceFeatures;
-        private AndroidSystem androidSystem;
-
-        private static Builder newInstance(Context context) {
-            Context applicationContext = context.getApplicationContext();
-            ExternalStorageDirectories externalStorageDirectories = new AndroidExternalStorageDirectories(applicationContext);
-            FileSystem fileSystem = new AndroidFileSystem(externalStorageDirectories);
-            DeviceFeatures deviceFeatures = new AndroidDeviceFeatures();
-            AndroidSystem androidSystem = new AndroidSystem();
-            return new Builder(
-                applicationContext,
-                fileSystem,
-                deviceFeatures,
-                androidSystem
-            );
-        }
-
-        private Builder(Context context, FileSystem fileSystem, DeviceFeatures deviceFeatures, AndroidSystem androidSystem) {
-            this.context = context;
-            this.fileSystem = fileSystem;
-            this.deviceFeatures = deviceFeatures;
-            this.androidSystem = androidSystem;
-        }
-
-        public Builder withFileSystem(FileSystem fileSystem) {
-            this.fileSystem = fileSystem;
-            return this;
-        }
-
-        public Builder withDeviceFeatures(DeviceFeatures deviceFeatures) {
-            this.deviceFeatures = deviceFeatures;
-            return this;
-        }
-
-        public Builder withAndroidSystem(AndroidSystem androidSystem) {
-            this.androidSystem = androidSystem;
-            return this;
-        }
-
-        public AndroidDeviceStorageInspector build() {
-            return new AndroidDeviceStorageInspector(
-                context,
-                fileSystem,
-                deviceFeatures,
-                androidSystem
-            );
-        }
     }
 }
